+++ ToolbarIcons/Framework/Models/Events/ToolbarIconsChangedEventArgs.cs
@@ -1,4 +1,0 @@
-namespace StardewMods.ToolbarIcons.Framework.Models.Events;
-
-/// <summary>Event raised when toolbar icons have been changed.</summary>
-internal sealed class ToolbarIconsChangedEventArgs : EventArgs { }